nav 구조 재수정중
