+++ tsconfig.json
@@ -2,7 +2,11 @@
   "compilerOptions": {
     "target": "es5",
     "downlevelIteration": true,
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -21,19 +25,22 @@
     ],
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"],
-      "@/icons/*": ["public/assets/icons/*"]
+      "@/*": [
+        "src/*"
+      ],
+      "@/icons/*": [
+        "public/assets/icons/*"
+      ]
     }
   },
   "include": [
     "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    "**/**/*.tsx",
-    "**/**/**/**/*.tsx",
-    "**/**/**/**/**/*.tsx",
-    ".next/types/**/*.ts",
-    "tailwind.config.js"
+    "src/**/*.tsx",
+    "src/**/*.ts",
+    "tailwind.config.js",
+    ".next/types/**/*.ts"
   ],
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
